mixed-content error fix try-7
