+++ src/components/tabs/data-management/procedure-management/bundle/bundle-create-modal.tsx
@@ -7,7 +7,8 @@ import { Textarea } from '@/components/ui/textarea';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { X, Save, Search, Package, Plus, Trash2 } from 'lucide-react';
 import { createBundle, BundleCreateRequest, BundleElementRequest } from '@/api/bundles-api';
-import { getElementsList, Element, searchElementsByName, getElementDetail } from '@/api/element-api';
+import { getElementsList, Element, getElementDetail } from '@/api/element-api';
+import { searchElementsByName } from '@/utils/element-utils';
 import { getConsumableDetail, ConsumableResponse } from '@/api/consumables-api';
 
 interface BundleCreateModalProps {
