+++ src/components/tabs/data-management/procedure-management/sequence/sequence-create-modal.tsx
@@ -6,7 +6,8 @@ import { Input } from '@/components/ui/input';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { X, Save, Plus, Trash2, Check } from 'lucide-react';
 import { SequenceCreateRequest, SequenceStepRequest, createSequence } from '@/api/sequences-api';
-import { Element, getElementsList, searchElementsByName } from '@/api/element-api';
+import { Element, getElementsList } from '@/api/element-api';
+import { searchElementsByName } from '@/utils/element-utils';
 import { BundleListResponse, getBundleDetail, getBundlesList, BundleElementResponse } from '@/api/bundles-api';
 import { CustomListResponse, getCustomDetail, getCustomsList, CustomElementResponse } from '@/api/customs-api';
 import { getConsumableDetail, ConsumableResponse } from '@/api/consumables-api';
