+++ src/components/tabs/data-management/procedure-management/sequence/sequence-detail-modal.tsx
@@ -6,7 +6,8 @@ import { Input } from '@/components/ui/input';
 import { X, Edit, Save, Plus, Tag, Trash2} from 'lucide-react';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/components/ui/accordion';
 import { SequenceResponse, SequenceStepResponse, updateSequence, deleteSequence } from '@/api/sequences-api';
-import { Element, getElementsList, searchElementsByName } from '@/api/element-api';
+import { Element, getElementsList } from '@/api/element-api';
+import { searchElementsByName } from '@/utils/element-utils';
 import { BundleListResponse, getBundlesList } from '@/api/bundles-api';
 import { CustomListResponse, getCustomsList } from '@/api/customs-api';
 import { ConsumableResponse } from '@/api/consumables-api';
