+++ src/components/tabs/data-management/procedure-management/sequence/sequnce-step-modal.tsx
@@ -4,7 +4,8 @@ import { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { X, Plus } from 'lucide-react';
-import { Element, getElementsList, searchElementsByName } from '@/api/element-api';
+import { Element, getElementsList } from '@/api/element-api';
+import { searchElementsByName } from '@/utils/element-utils';
 import { BundleListResponse, getBundlesList, getBundleDetail } from '@/api/bundles-api';
 import { CustomListResponse, getCustomsList, getCustomDetail } from '@/api/customs-api';
 import { ConsumableResponse, getConsumableDetail } from '@/api/consumables-api';
